api: pipeline refactor
* removed pipeline base class: the only common generic methods `predict`
  will not be kept for new workflows. It makes more sense to have
  something like `get_OUTPUT` where OUTPUT will match the headless mode.
* `PixelClassificationPipeline` supersedes the previous hidden
  `_PipelineImpl` class with a `from_project_file` method for creating
  such objects -> clear what class is returned.
