+++ ilastik/experimental/api/__init__.py
@@ -1,2 +1,1 @@
-from ._base import from_project_file
-from .types import Pipeline
+from ._pipelines import PixelClassificationPipeline, from_project_file
