--- ilastik/experimental/api/types.py
@@ -1,8 +1,0 @@
-import numpy
-import abc
-
-
-class Pipeline(abc.ABC):
-    @abc.abstractmethod
-    def predict(self, data: numpy.ndarray) -> numpy.ndarray:
-        ...
